Move functions to their own translation unit and improve docstrings
