+++ CMakeLists.txt
@@ -28,6 +28,8 @@ add_executable(fcc_bridge
         src/fcc_bridge_node_mavsdk.cpp
         src/fcc_bridge_node_conversion.cpp
         src/fcc_bridge_node_safety.cpp
+        src/fcc_bridge_node_commands.cpp
+        src/fcc_bridge_node_telemetry.cpp
 )
 
 target_link_libraries(fcc_bridge
