+++ src/fcc_bridge_node_ros.cpp
@@ -46,9 +46,6 @@ constexpr char const *const UAV_COMMAND_TOPIC_NAME =
 constexpr std::chrono::seconds MAX_UAV_COMMAND_AGE{
     1}; /**< Maximum age of a received UAV command */
 
-constexpr float WAYPOINT_ACCEPTANCE_RADIUS_M =
-    .5; /**< Acceptance radius for waypoints */
-
 }  // namespace
 
 void FCCBridgeNode::set_internal_state(
@@ -269,300 +266,6 @@ void FCCBridgeNode::uav_command_subscriber_cb(
     RCLCPP_DEBUG(this->get_logger(), "Finished handling command");
 }
 
-void FCCBridgeNode::initiate_takeoff(const interfaces::msg::Waypoint &waypoint,
-                                     const float speed_mps) {
-    RCLCPP_INFO(this->get_logger(),
-                "Received a command to takeoff to lat: %f°\tlon: %f°\trel alt: "
-                "%fm with speed: %fm/s",
-                waypoint.latitude_deg, waypoint.longitude_deg,
-                waypoint.relative_altitude_m, static_cast<double>(speed_mps));
-
-    // Verify that the uav is in the right state
-    switch (this->get_internal_state()) {
-        case INTERNAL_STATE::ERROR:
-            // This should never happen as the process should have exited before
-            throw std::runtime_error(
-                "Received command to take of while in ERROR state");
-        case INTERNAL_STATE::STARTING_UP:
-        case INTERNAL_STATE::ROS_SET_UP:
-        case INTERNAL_STATE::MAVSDK_SET_UP:
-        case INTERNAL_STATE::WAITING_FOR_ARM:
-        case INTERNAL_STATE::LANDED:
-            // The UAV is still or again on the ground so the only thing left is
-            // to kill the process
-            RCLCPP_FATAL(
-                this->get_logger(),
-                "Received a takeoff command in an invalid state! Exiting...");
-            this->set_internal_state(INTERNAL_STATE::ERROR);
-            this->exit_process_on_error();
-        case INTERNAL_STATE::WAITING_FOR_COMMAND:
-        case INTERNAL_STATE::FLYING_ACTION:
-        case INTERNAL_STATE::FLYING_MISSION:
-        case INTERNAL_STATE::RETURN_TO_HOME:
-            // If the UAV is airborne and another takeoff command is received
-            // something must have gone wrong so an RTH is triggered
-            RCLCPP_ERROR(this->get_logger(),
-                         "Received a takeoff command while airborne! "
-                         "Triggering RTH...");
-            this->trigger_rth();
-            return;
-        case INTERNAL_STATE::ARMED:
-            break;
-        default:
-            throw std::runtime_error(
-                std::string("Got invalid value for internal_state: ") +
-                std::to_string(static_cast<int>(this->get_internal_state())));
-    }
-
-    // Check that the speed is valid
-    if (!this->check_speed(speed_mps)) {
-        if (this->get_internal_state() == INTERNAL_STATE::ERROR) {
-            // This means an unrecoverable error has occurred such as missing
-            // safety limits
-            RCLCPP_FATAL(
-                this->get_logger(),
-                "There has been an unrecoverable error checking if the target "
-                "speed for takeoff was inside the safety limits! Exiting...");
-        } else {
-            // In this case the target takeoff speed is outside the limits
-            // deemed safe. Because the UAV is on the ground this results in a
-            // process exit.
-            RCLCPP_FATAL(
-                this->get_logger(),
-                "The target takeoff speed is outside the speed "
-                "safety limits (%f is not in (%f;%f]). The UAV should "
-                "not be airborne! Exiting...",
-                static_cast<double>(speed_mps),
-                static_cast<double>(safety_limits::MIN_SPEED_LIMIT_MPS),
-                static_cast<double>(this->safety_limits->max_speed_mps));
-            this->set_internal_state(INTERNAL_STATE::ERROR);
-        }
-        // Exit the process
-        this->exit_process_on_error();
-    }
-
-    // Check that the waypoint is valid
-    if (waypoint.relative_altitude_m ==
-        interfaces::msg::Waypoint::INVALID_ALTITUDE) {
-        RCLCPP_FATAL(
-            this->get_logger(),
-            "Got an invalid waypoint for a takeoff command! Exiting...");
-        this->set_internal_state(INTERNAL_STATE::ERROR);
-        this->exit_process_on_error();
-    }
-
-    // Verify that the waypoint is inside the geofence
-    if (!this->check_point_in_geofence(waypoint)) {
-        // Something went wrong with the check
-        if (this->get_internal_state() == INTERNAL_STATE::ERROR) {
-            // This means an unrecoverable error has occurred such as missing
-            // safety limits
-            RCLCPP_FATAL(
-                this->get_logger(),
-                "There has been an unrecoverable error checking if the target "
-                "waypoint for takeoff was inside the geofence! Exiting...");
-        } else {
-            // The waypoint is outside the geofence. The UAV is still on the
-            // ground so the process will exit.
-            RCLCPP_FATAL(
-                this->get_logger(),
-                "Got an waypoint with lat: %f°\tlon: %f°\trel. alt. %fm "
-                "for a takeoff that is outside the geofence! Exiting...",
-                waypoint.latitude_deg, waypoint.longitude_deg,
-                waypoint.relative_altitude_m);
-            this->set_internal_state(INTERNAL_STATE::ERROR);
-        }
-        // Exit the process
-        this->exit_process_on_error();
-    }
-
-    RCLCPP_INFO(this->get_logger(), "Clear for takeoff");
-
-    // Mission item that describes the takeoff action
-    mavsdk::Mission::MissionItem takeoff_mission_item;
-
-    // Set all the relevant values
-    takeoff_mission_item.latitude_deg = waypoint.latitude_deg;
-    takeoff_mission_item.longitude_deg = waypoint.longitude_deg;
-    takeoff_mission_item.relative_altitude_m = waypoint.relative_altitude_m;
-    takeoff_mission_item.speed_m_s = speed_mps;
-    takeoff_mission_item.is_fly_through = false;
-    takeoff_mission_item.acceptance_radius_m = WAYPOINT_ACCEPTANCE_RADIUS_M;
-    takeoff_mission_item.vehicle_action =
-        mavsdk::Mission::MissionItem::VehicleAction::Takeoff;
-
-    // Mission plan that stores the mission items. In this case only one
-    mavsdk::Mission::MissionPlan takeoff_mission_plan;
-
-    // Insert takeoff action into mission plan
-    takeoff_mission_plan.mission_items.push_back(takeoff_mission_item);
-
-    // Execute mission
-    if (!this->execute_mission_plan(takeoff_mission_plan)) {
-        RCLCPP_FATAL(
-            this->get_logger(),
-                     "There was an error triggering the waypoint mission! "
-                     "Triggering RTH...");
-        this->trigger_rth();
-        return;
-    }
-
-    // Set state to waiting for mission
-    this->set_internal_state(INTERNAL_STATE::FLYING_MISSION);
-
-    RCLCPP_INFO(
-        this->get_logger(),
-        "Successfully triggered takeoff. Waiting for mission to complete.");
-}
-
-void FCCBridgeNode::start_flying_to_waypoint(
-    const interfaces::msg::Waypoint &waypoint, const float speed_mps) {
-    RCLCPP_INFO(this->get_logger(),
-                "Received a command to fly to a waypoint at lat: %f°\tlon: "
-                "%f°\trel alt: %fm with speed: %fm/s",
-                waypoint.latitude_deg, waypoint.longitude_deg,
-                waypoint.relative_altitude_m, static_cast<double>(speed_mps));
-
-    // Verify that the uav is in the right state
-    switch (this->get_internal_state()) {
-        case INTERNAL_STATE::ERROR:
-            // This should never happen as the process should have exited before
-            throw std::runtime_error(
-                "Received command to take of while in ERROR state");
-        case INTERNAL_STATE::STARTING_UP:
-        case INTERNAL_STATE::ROS_SET_UP:
-        case INTERNAL_STATE::MAVSDK_SET_UP:
-        case INTERNAL_STATE::WAITING_FOR_ARM:
-        case INTERNAL_STATE::LANDED:
-        case INTERNAL_STATE::ARMED:
-            // The UAV is still or again on the ground so the only thing left is
-            // to kill the process
-            RCLCPP_FATAL(
-                this->get_logger(),
-                "Received a takeoff command in an invalid state! Exiting...");
-            this->set_internal_state(INTERNAL_STATE::ERROR);
-            this->exit_process_on_error();
-        case INTERNAL_STATE::FLYING_ACTION:
-        case INTERNAL_STATE::FLYING_MISSION:
-        case INTERNAL_STATE::RETURN_TO_HOME:
-            // If the UAV is already waiting on a UAVCommand and another
-            // waypoint command is received something must have gone wrong so an
-            // RTH is triggered
-            RCLCPP_ERROR(this->get_logger(),
-                         "Received a waypoint command while there is already a "
-                         "command in progress! Triggering RTH...");
-            this->trigger_rth();
-            return;
-        case INTERNAL_STATE::WAITING_FOR_COMMAND:
-            break;
-        default:
-            throw std::runtime_error(
-                std::string("Got invalid value for internal_state: ") +
-                std::to_string(static_cast<int>(this->get_internal_state())));
-    }
-
-    // Check that the speed is valid
-    if (!this->check_speed(speed_mps)) {
-        if (this->get_internal_state() == INTERNAL_STATE::ERROR) {
-            // This means an unrecoverable error has occurred such as missing
-            // safety limits
-            RCLCPP_FATAL(
-                this->get_logger(),
-                "There has been an unrecoverable error checking if the target "
-                "speed for waypoint was inside the safety limits! Exiting...");
-            // Exit the process
-            this->exit_process_on_error();
-        } else {
-            // In this case the target waypoint speed is outside the limits
-            // deemed safe. This will result in an RTH
-            RCLCPP_ERROR(
-                this->get_logger(),
-                "The target waypoint speed is outside the speed safety limits "
-                "(%f is not in (%f;%f]). Triggering RTH...",
-                static_cast<double>(speed_mps),
-                static_cast<double>(safety_limits::MIN_SPEED_LIMIT_MPS),
-                static_cast<double>(this->safety_limits->max_speed_mps));
-            this->trigger_rth();
-            return;
-        }
-    }
-
-    // Check that the waypoint is valid
-    if (waypoint.relative_altitude_m ==
-        interfaces::msg::Waypoint::INVALID_ALTITUDE) {
-        RCLCPP_ERROR(
-            this->get_logger(),
-            "Got an invalid waypoint for a takeoff command! Triggering RTH...");
-        this->trigger_rth();
-        return;
-    }
-
-    // Verify that the waypoint is inside the geofence
-    if (!this->check_point_in_geofence(waypoint)) {
-        // Something went wrong with the check
-        if (this->get_internal_state() == INTERNAL_STATE::ERROR) {
-            // This means an unrecoverable error has occurred such as missing
-            // safety limits
-            RCLCPP_FATAL(
-                this->get_logger(),
-                "There has been an unrecoverable error checking if the target "
-                "waypoint was inside the geofence! Exiting...");
-            // Exit the process
-            this->exit_process_on_error();
-        } else {
-            // The waypoint is outside the geofence. This will trigger an RTH.
-            RCLCPP_FATAL(
-                this->get_logger(),
-                "Got an waypoint with lat: %f°\tlon: %f°\trel. alt. %fm to fly "
-                "to that is outside the geofence! Triggering RTH...",
-                waypoint.latitude_deg, waypoint.longitude_deg,
-                waypoint.relative_altitude_m);
-            this->trigger_rth();
-            return;
-        }
-    }
-
-    RCLCPP_INFO(this->get_logger(), "Clear for fly to waypoint");
-
-    // Mission item that describes the fly to waypoint action
-    mavsdk::Mission::MissionItem fly_to_waypoint_mission_item;
-
-    // Set all the relevant values
-    fly_to_waypoint_mission_item.latitude_deg = waypoint.latitude_deg;
-    fly_to_waypoint_mission_item.longitude_deg = waypoint.longitude_deg;
-    fly_to_waypoint_mission_item.relative_altitude_m =
-        fly_to_waypoint_mission_item.relative_altitude_m;
-    fly_to_waypoint_mission_item.speed_m_s = speed_mps;
-    fly_to_waypoint_mission_item.is_fly_through = false;
-    fly_to_waypoint_mission_item.acceptance_radius_m =
-        WAYPOINT_ACCEPTANCE_RADIUS_M;
-    fly_to_waypoint_mission_item.vehicle_action =
-        mavsdk::Mission::MissionItem::VehicleAction::None;
-
-    // Mission plan that stores the mission items. In this case only one
-    mavsdk::Mission::MissionPlan fly_to_waypoint_mission_plan;
-
-    // Insert fly to waypoint action into mission plan
-    fly_to_waypoint_mission_plan.mission_items.push_back(
-        fly_to_waypoint_mission_item);
-
-    // Execute mission
-    if (!this->execute_mission_plan(fly_to_waypoint_mission_plan)) {
-        RCLCPP_FATAL(
-            this->get_logger(),
-            "There was an error triggering the takeoff mission! Exiting...");
-        this->set_internal_state(INTERNAL_STATE::ERROR);
-        this->exit_process_on_error();
-    }
-
-    // Set state to waiting for mission
-    this->set_internal_state(INTERNAL_STATE::FLYING_MISSION);
-
-    RCLCPP_INFO(this->get_logger(),
-                "Successfully triggered fly to waypoint. Waiting for mission "
-                "to complete.");
-}
-
 void FCCBridgeNode::fcc_telemetry_timer_5hz_cb() {
     RCLCPP_DEBUG(this->get_logger(),
                  "5Hz telemetry timer callback was triggered");
@@ -648,233 +351,6 @@ void FCCBridgeNode::fcc_telemetry_timer_10hz_cb() {
 
     RCLCPP_DEBUG(this->get_logger(),
                  "10Hz telemetry timer callback successful");
-}
-
-void FCCBridgeNode::check_last_mission_control_heartbeat() {
-    // TODO: Check if last heartbeat is not too old.
-}
-
-void FCCBridgeNode::send_gps_telemetry() {
-    RCLCPP_DEBUG(this->get_logger(),
-                 "Getting updated GPS telemetry and publishing the update");
-
-    this->get_gps_telemetry();
-
-    // Verify the GPS Fix type depending on the internal state
-    this->check_gps_state();
-
-    // In this case retrieving the gps telemetry was successful meaning that it
-    // can be safely accessed.
-    interfaces::msg::GPSPosition gps_msg;
-    gps_msg.time_stamp = this->now();
-    gps_msg.sender_id = this->get_name();
-    gps_msg.fix_type =
-        FCCBridgeNode::fix_type_mavsdk_to_ros(last_fcc_gps_info->fix_type);
-    gps_msg.latitude_deg = last_fcc_position->latitude_deg;
-    gps_msg.longitude_deg = last_fcc_position->longitude_deg;
-    gps_msg.relative_altitude_m = last_fcc_position->relative_altitude_m;
-
-    // Publish the message
-    this->gps_position_publisher->publish(gps_msg);
-
-    RCLCPP_DEBUG(this->get_logger(), "Published current GPS position");
-}
-
-void FCCBridgeNode::send_flight_state() {
-    RCLCPP_DEBUG(this->get_logger(),
-                 "Getting updated flight state and publishing the update");
-
-    // Update flight state
-    this->get_flight_state();
-
-    // Verify the Flight State depending on the internal state
-    this->check_flight_state();
-
-    // If we are in the WAITING_FOR_ARM state and the drone is READY send
-    // MissionStart and wait for takeoff
-    if (this->get_internal_state() == INTERNAL_STATE::WAITING_FOR_ARM &&
-        this->last_fcc_flight_mode.value() ==
-            mavsdk::Telemetry::FlightMode::Ready) {
-        RCLCPP_INFO(this->get_logger(),
-                    "Drone is armed and ready for takeoff. Switching to ARMED "
-                    "state and sending MissionStart message");
-        this->set_internal_state(INTERNAL_STATE::ARMED);
-        interfaces::msg::MissionStart mission_start_msg;
-        mission_start_msg.sender_id = this->get_name();
-        this->mission_start_publisher->publish(mission_start_msg);
-        RCLCPP_DEBUG(this->get_logger(), "MissionStart message send");
-    }
-
-    // In this case retrieving the flight state was successful meaning that it
-    // can be safely accessed.
-    interfaces::msg::FlightState flight_state_msg;
-    flight_state_msg.time_stamp = this->now();
-    flight_state_msg.sender_id = this->get_name();
-    flight_state_msg.mode.mode = FCCBridgeNode::flight_mode_mavsdk_to_ros(
-        this->last_fcc_flight_mode.value());
-    flight_state_msg.state.state = FCCBridgeNode::landed_state_mavsdk_to_ros(
-        this->last_fcc_landed_state.value());
-
-    // Publish the message
-    this->flight_state_publisher->publish(flight_state_msg);
-
-    RCLCPP_DEBUG(this->get_logger(), "Published current flight state");
-}
-
-void FCCBridgeNode::send_battery_state() {
-    RCLCPP_DEBUG(this->get_logger(),
-                 "Getting updated battery state and publishing the update");
-
-    // Update battery state
-    this->get_battery_state();
-
-    // Verify the Battery State depending on the internal state
-    this->check_battery_state();
-
-    // In this case retrieving the battery state was successful meaning that it
-    // can be safely accessed.
-    interfaces::msg::BatteryState battery_state_msg;
-    battery_state_msg.time_stamp = this->now();
-    battery_state_msg.sender_id = this->get_name();
-    battery_state_msg.id = this->last_fcc_battery_state->id;
-    battery_state_msg.temperature_degc =
-        this->last_fcc_battery_state->temperature_degc;
-    battery_state_msg.voltage_v = this->last_fcc_battery_state->voltage_v;
-    battery_state_msg.current_battery_a =
-        this->last_fcc_battery_state->current_battery_a;
-    battery_state_msg.capacity_consumed_ah =
-        this->last_fcc_battery_state->capacity_consumed_ah;
-    battery_state_msg.remaining_percent =
-        this->last_fcc_battery_state->remaining_percent;
-
-    // Publish the message
-    this->battery_state_publisher->publish(battery_state_msg);
-
-    RCLCPP_DEBUG(this->get_logger(), "Published current battery state");
-}
-
-void FCCBridgeNode::send_rc_state() {
-    RCLCPP_DEBUG(this->get_logger(),
-                 "Getting updated RC state and publishing the update");
-
-    // Update RC state
-    this->get_rc_state();
-
-    // Verify the RC State depending on the internal state
-    this->check_rc_state();
-
-    // In this case retrieving the RC state was successful meaning that it
-    // can be safely accessed.
-    interfaces::msg::RCState rc_state_msg;
-    rc_state_msg.time_stamp = this->now();
-    rc_state_msg.sender_id = this->get_name();
-    rc_state_msg.was_available_once =
-        this->last_fcc_rc_state->was_available_once;
-    rc_state_msg.is_available = this->last_fcc_rc_state->is_available;
-    rc_state_msg.signal_strength_percent =
-        this->last_fcc_rc_state->signal_strength_percent;
-
-    // Publish the message
-    this->rc_state_publisher->publish(rc_state_msg);
-
-    RCLCPP_DEBUG(this->get_logger(), "Published current RC state");
-}
-
-void FCCBridgeNode::send_euler_angle() {
-    RCLCPP_DEBUG(this->get_logger(),
-                 "Getting updated euler angle and publishing the update");
-
-    // Update euler angle
-    this->get_euler_angle();
-
-    // In this case retrieving the euler angle was successful meaning that it
-    // can be safely accessed.
-    interfaces::msg::Pose euler_angle_msg;
-    euler_angle_msg.time_stamp = this->now();
-    euler_angle_msg.sender_id = this->get_name();
-    euler_angle_msg.roll_deg = this->last_fcc_euler_angle->roll_deg;
-    euler_angle_msg.pitch_deg = this->last_fcc_euler_angle->pitch_deg;
-    euler_angle_msg.yaw_deg = this->last_fcc_euler_angle->yaw_deg;
-
-    // Publish the message
-    this->euler_angle_publisher->publish(euler_angle_msg);
-
-    RCLCPP_DEBUG(this->get_logger(), "Published current euler angle");
-}
-
-void FCCBridgeNode::send_mission_progress() {
-    RCLCPP_DEBUG(this->get_logger(),
-                 "Getting updated mission progress and publishing the update");
-
-    // Update mission progress
-    this->get_mission_progress();
-
-    // Verify of getting the mission progress was successful
-    if (this->last_mission_progress->first !=
-        mavsdk::Mission::Result::Success) {
-        // Trigger an RTH on any error
-        RCLCPP_ERROR(this->get_logger(),
-                     "Failed to get mission progress! Triggering RTH...");
-        this->trigger_rth();
-        return;
-    }
-
-    // In this case retrieving the mission progress was successful meaning that
-    // it can be safely accessed.
-    interfaces::msg::MissionProgress mission_progress_msg;
-    mission_progress_msg.time_stamp = this->now();
-    mission_progress_msg.sender_id = this->get_name();
-    // Check if the mission is finished
-    if (this->last_mission_progress->second) {
-        // Switch to the waiting for command state
-        this->set_internal_state(INTERNAL_STATE::WAITING_FOR_COMMAND);
-        mission_progress_msg.progress = 1.0;
-    } else {
-        mission_progress_msg.progress = 0.0;
-    }
-
-    // Publish the message
-    this->mission_progress_publisher->publish(mission_progress_msg);
-
-    RCLCPP_DEBUG(this->get_logger(), "Published current mission progress");
-}
-
-void FCCBridgeNode::send_uav_health() {
-    RCLCPP_DEBUG(this->get_logger(),
-                 "Getting updated uav health and publishing the update");
-
-    // Update UAV health
-    this->get_uav_health();
-
-    // Verify the health depending on the internal state
-    this->check_uav_health();
-
-    RCLCPP_INFO(this->get_logger(), "UAV config is ok in the current state %s",
-                this->internal_state_to_str());
-
-    // In this case retrieving the UAV health was successful meaning that it
-    // can be safely accessed.
-    interfaces::msg::UAVHealth uav_health_msg;
-    uav_health_msg.time_stamp = this->now();
-    uav_health_msg.sender_id = this->get_name();
-    uav_health_msg.is_gyrometer_calibration_ok =
-        this->last_fcc_health->is_gyrometer_calibration_ok;
-    uav_health_msg.is_accelerometer_calibration_ok =
-        this->last_fcc_health->is_accelerometer_calibration_ok;
-    uav_health_msg.is_magnetometer_calibration_ok =
-        this->last_fcc_health->is_magnetometer_calibration_ok;
-    uav_health_msg.is_local_position_ok =
-        this->last_fcc_health->is_local_position_ok;
-    uav_health_msg.is_global_position_ok =
-        this->last_fcc_health->is_global_position_ok;
-    uav_health_msg.is_home_position_ok =
-        this->last_fcc_health->is_home_position_ok;
-    uav_health_msg.is_armable = this->last_fcc_health->is_armable;
-
-    // Publish the message
-    this->uav_health_publisher->publish(uav_health_msg);
-
-    RCLCPP_DEBUG(this->get_logger(), "Published current UAV health");
 }
 
 FCCBridgeNode::FCCBridgeNode(const std::string &name/*,
